phase 2 WIP..., we have implemented the memory pool and intrusiev list to prevent memort allocation in hot path. The add_limit_order() and add_market_order() functions has been adapted. What left? 1. modify the cancel_order() function; 2. run 7 scenario benchmark, plot the performance improvement; 3. implement the jump table to replace the std::map, achieve ultimate O(1) add/delete/check/modify

+++ benchmark/src/bench_common.hpp
@@ -128,7 +128,7 @@ class PerfGroup {
    * @param group_fd Leader FD, or -1 to make this counter the group leader.
    * @return true if the counter was opened and added to the group.
    */
-  bool OpenCounter(std::uint32_t type, std::uint64_t config, int group_fd) {
+  bool OpenCounter(std::uint64_t type, std::uint64_t config, int group_fd) {
     struct perf_event_attr pe {};
     pe.type = type;
     pe.size = sizeof(struct perf_event_attr);
